refactor: Session 관리 UserRepository로 통합
